refactor(e2e): split flat tests/ directory into feature subfolders
Organizes the 11 spec files into account/, auth/, dashboard/, layer/,
project/ by feature area. Each move updates only its relative import
paths (one extra ../ for the added directory depth) -- no test logic
changed. Verified via --list: still resolves to the same 150 tests
across all files, no losses or duplicates.
