+++ e2e/tests/account/accountWorkspaceSettings.spec.ts
@@ -1,10 +1,10 @@
 import { test, expect, BrowserContext, Page } from "@playwright/test";
 
-import { STORAGE_STATE } from "../global-setup";
-import { AccountSettingsPage } from "../pages/accountSettingsPage";
-import { DashBoardPage } from "../pages/dashBoardPage";
-import { WorkspaceSettingsPage } from "../pages/workspaceSettingsPage";
-import { createIAPContext } from "../utils/iap-auth";
+import { STORAGE_STATE } from "../../global-setup";
+import { AccountSettingsPage } from "../../pages/accountSettingsPage";
+import { DashBoardPage } from "../../pages/dashBoardPage";
+import { WorkspaceSettingsPage } from "../../pages/workspaceSettingsPage";
+import { createIAPContext } from "../../utils/iap-auth";
 
 const REEARTH_WEB_E2E_BASEURL = process.env.REEARTH_WEB_E2E_BASEURL;
 if (!REEARTH_WEB_E2E_BASEURL) {
