+++ e2e/tests/account/members.spec.ts
@@ -1,9 +1,9 @@
 import { test, expect, BrowserContext, Page } from "@playwright/test";
 
-import { STORAGE_STATE } from "../global-setup";
-import { DashBoardPage } from "../pages/dashBoardPage";
-import { MembersPage } from "../pages/membersPage";
-import { createIAPContext } from "../utils/iap-auth";
+import { STORAGE_STATE } from "../../global-setup";
+import { DashBoardPage } from "../../pages/dashBoardPage";
+import { MembersPage } from "../../pages/membersPage";
+import { createIAPContext } from "../../utils/iap-auth";
 
 const REEARTH_WEB_E2E_BASEURL = process.env.REEARTH_WEB_E2E_BASEURL;
 if (!REEARTH_WEB_E2E_BASEURL) {
