+++ e2e/tests/layer/multipleStyles.spec.ts
@@ -1,14 +1,14 @@
 import { faker } from "@faker-js/faker";
 import { test, expect, BrowserContext, Page } from "@playwright/test";
 
-import { STORAGE_STATE } from "../global-setup";
-import { CesiumViewerPage } from "../pages/cesiumViewerPage";
-import { DashBoardPage } from "../pages/dashBoardPage";
-import { LayerStylePanelPage } from "../pages/layerStylePanelPage";
-import { ProjectScreenPage } from "../pages/projectScreenPage";
-import { ProjectsPage } from "../pages/projectsPage";
-import { createIAPContext } from "../utils/iap-auth";
-import { deleteProjectByName } from "../utils/project-cleanup";
+import { STORAGE_STATE } from "../../global-setup";
+import { CesiumViewerPage } from "../../pages/cesiumViewerPage";
+import { DashBoardPage } from "../../pages/dashBoardPage";
+import { LayerStylePanelPage } from "../../pages/layerStylePanelPage";
+import { ProjectScreenPage } from "../../pages/projectScreenPage";
+import { ProjectsPage } from "../../pages/projectsPage";
+import { createIAPContext } from "../../utils/iap-auth";
+import { deleteProjectByName } from "../../utils/project-cleanup";
 
 const REEARTH_E2E_EMAIL = process.env.REEARTH_E2E_EMAIL;
 const REEARTH_E2E_PASSWORD = process.env.REEARTH_E2E_PASSWORD;
